Split chapter entity tests
